requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
influxdb-client>=1.38.0
python-snap7>=1.3
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0
psutil>=5.9.0
pytest>=7.4.0
httpx>=0.25.0
pymodbus>=3.6.0
pyserial>=3.5
